requirements.txt
pygame
glfw
pyopengl
imgui
python-socketio
pysdl2